+++ test/contracts/access/SafeContractOwnership.test.js
@@ -205,6 +205,6 @@ runBehaviorTests('SafeContractOwnership', config, function (deployFn) {
       });
     });
 
-    supportsInterfaces(['IERC165', 'IERC173']);
+    supportsInterfaces(['@openzeppelin/contracts/utils/introspection/IERC165.sol:IERC165', 'IERC173']);
   });
 });