+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Updated `ForwarderRegistry` so that forwarder approvals are set per meta-tx target contract instead of globally.
 - Replaced `ForwarderRegistry`.`removeForwarderApproval` function with `ForwarderRegistry`.`setForwarderApproval`.
 - Upgraded to `solc@0.8.25`.
-- Upgraded to `openzeppelin/contracts@5.0.2`.
+- Upgraded to `openzeppelin/contracts@5.2.0`.
 - Upgraded to new Node version.
 
 ### New features